+++ client/src/showAndCopyOtpCode.ts
@@ -1,4 +1,4 @@
-import { client } from '../api'
+import { client } from './api'
 
 export async function showAndCopyOtpCode(
 	id: string,
